rename 'ModuleHeader.vue' to 'ModuleProgressBar.vue'
