+++ resources/js/components/modules/ModuleProgressBar.vue
